Refactor: update syntax of provider requirements
TF 0.13 introduced this new syntax where each provider is declared as a
map with a source and version (contrasting with how things were before,
where you just set the version you wanted). We should always use the new
syntax for explicitness and because the pre-0.13 syntax may be
deprecated some day.