Manage the cluster's outbound IP ourselves
By default, when not setting a public IP when creating an AKS cluster,
one is created automatically. Instead of doing that, we'll now create
the public IP ourselves and tell AKS to use it. Currently we'll only
support a single IP address as it's easier. This also allows us to
export that IP address in the module's outputs, which is useful when a
cluster will need to connect to another endpoint that has an allowlist.
This makes it easier to manage that allowlist via code.